add aware package
