+++ sharding-core/sharding-core-rewrite/src/main/java/org/apache/shardingsphere/core/rewrite/feature/sharding/parameter/ShardingParameterRewriterBuilder.java
@@ -21,8 +21,8 @@ import lombok.RequiredArgsConstructor;
 import org.apache.shardingsphere.core.metadata.table.TableMetas;
 import org.apache.shardingsphere.core.rewrite.feature.sharding.parameter.impl.ShardingGeneratedKeyInsertValueParameterRewriter;
 import org.apache.shardingsphere.core.rewrite.feature.sharding.parameter.impl.ShardingPaginationParameterRewriter;
-import org.apache.shardingsphere.core.rewrite.feature.sharding.token.generator.SQLRouteResultAware;
-import org.apache.shardingsphere.core.rewrite.feature.sharding.token.generator.ShardingRuleAware;
+import org.apache.shardingsphere.core.rewrite.feature.sharding.aware.SQLRouteResultAware;
+import org.apache.shardingsphere.core.rewrite.feature.sharding.aware.ShardingRuleAware;
 import org.apache.shardingsphere.core.rewrite.parameter.rewriter.ParameterRewriter;
 import org.apache.shardingsphere.core.rewrite.parameter.rewriter.ParameterRewriterBuilder;
 import org.apache.shardingsphere.core.rewrite.sql.token.generator.aware.TableMetasAware;
